processes can now be cancelled via HTTP request. see updated postman collection.
postman collection has examples of new ways to DELETE processes as well as more options in the listings we can GET about processes. this HTTP-based DELETE is a safer means of cancelling out of a training process and should likely be run before cancelling the node process running the mcts server, since faulty data in the neo4j database can occur when exiting while training is running.